backend/pyproject.toml
[tool.poetry]
name = "random-dish"
version = "0.1.0"
description = ""
authors = ["shin-hama <htysivista@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.63.0"}
python-dotenv = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.6"
flake8 = "^3.9.0"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
